financial_image_scans/requirements.txt
streamlit
Pillow
requests
google-generativeai
python-dotenv
